requirements.txt
flask
flask_cors
pytesseract
pillow